Fixed the timing of setting some game mode flags
* now WORLDMAP, PIPBOY, INVENTORY, INTFACEUSE, and INTFACELOOT are
toggled after creating/before deleting their corresponding interface
windows.

Fixed the array out of bounds bug in Skills.cpp.

+++ sfall/Stats.cpp
@@ -207,7 +207,7 @@ allow:
 }
 
 static void StatsReset() {
-	for (int i = 0; i < STAT_max_stat; i++) {
+	for (size_t i = 0; i < STAT_max_stat; i++) {
 		StatMaximumsPC[i] = StatMaximumsNPC[i] = *(DWORD*)(_stat_data + 16 + i * 24);
 		StatMinimumsPC[i] = StatMinimumsNPC[i] = *(DWORD*)(_stat_data + 12 + i * 24);
 	}
